♻️ refactor: reorganize CLI into cmd/anna and config packages
- Move all CLI commands from root to cmd/anna/ following Go standard layout
- Extract config types and loading logic into dedicated config/ package
- Split monolithic main.go into focused command files:
  - main.go: minimal entry point
  - commands.go: app setup and factories
  - chat.go, gateway.go, models.go, skills.go, version.go, onboard.go
- Update mise.toml and .goreleaser.yaml build paths
- Maintain all existing functionality and tests
