+++ config/config.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 import (
 	"fmt"
@@ -155,9 +155,9 @@ type TelegramConfig struct {
 func (c TelegramConfig) IsEnabled() bool       { return c.Enabled == nil || *c.Enabled }
 func (c TelegramConfig) IsNotifyEnabled() bool { return c.EnableNotify != nil && *c.EnableNotify }
 
-// annaHome returns the anna home directory.
-// Priority: ANNA_HOME env → ~/.anna
-func annaHome() string {
+// AnnaHome returns the anna home directory.
+// Priority: ANNA_HOME env -> ~/.anna
+func AnnaHome() string {
 	if v := os.Getenv("ANNA_HOME"); v != "" {
 		return v
 	}
@@ -168,9 +168,9 @@ func annaHome() string {
 	return filepath.Join(home, ".anna")
 }
 
-// configPath returns the path to config.yaml inside the anna home.
-func configPath() string {
-	return filepath.Join(annaHome(), "config.yaml")
+// Path returns the path to config.yaml inside the anna home.
+func Path() string {
+	return filepath.Join(AnnaHome(), "config.yaml")
 }
 
 // StatePath returns the path to state.yaml (mutable runtime state) inside the workspace.
@@ -178,18 +178,18 @@ func (cfg *Config) StatePath() string {
 	return filepath.Join(cfg.Workspace, "state.yaml")
 }
 
-// cachePath returns the cache directory inside the anna home.
-func cachePath() string {
-	return filepath.Join(annaHome(), "cache")
+// CachePath returns the cache directory inside the anna home.
+func CachePath() string {
+	return filepath.Join(AnnaHome(), "cache")
 }
 
-// LoadConfig loads config from the default anna home (~/.anna/config.yaml).
-func LoadConfig() (*Config, error) {
-	return loadConfigFrom(annaHome())
+// Load loads config from the default anna home (~/.anna/config.yaml).
+func Load() (*Config, error) {
+	return LoadFrom(AnnaHome())
 }
 
-// loadConfigFrom loads config from the given directory.
-func loadConfigFrom(dir string) (*Config, error) {
+// LoadFrom loads config from the given directory.
+func LoadFrom(dir string) (*Config, error) {
 	cfg := &Config{}
 
 	if err := os.MkdirAll(dir, 0o755); err != nil {
@@ -207,7 +207,7 @@ func loadConfigFrom(dir string) (*Config, error) {
 	}
 
 	// Resolve workspace early so state.yaml can be found.
-	// Priority: ANNA_WORKSPACE env → yaml → default.
+	// Priority: ANNA_WORKSPACE env -> yaml -> default.
 	if cfg.Workspace == "" {
 		if v := os.Getenv("ANNA_WORKSPACE"); v != "" {
 			cfg.Workspace = v
@@ -289,7 +289,7 @@ func (cfg *Config) SkillsPath() string {
 }
 
 func (cfg *Config) ModelsPath() string {
-	return filepath.Join(cachePath(), "models.json")
+	return filepath.Join(CachePath(), "models.json")
 }
 
 func (cfg *Config) LogPath() string {
@@ -303,9 +303,9 @@ func (cfg *Config) ResolveModel() types.Model {
 }
 
 // ResolveModelTier returns the model for the given tier after applying
-// the fallback: strong → model, fast → model.
+// the fallback: strong -> model, fast -> model.
 func (cfg *Config) ResolveModelTier(tier string) types.Model {
-	modelID := cfg.resolveModelID(tier)
+	modelID := cfg.ResolveModelID(tier)
 	providerCfg := cfg.Providers[cfg.Provider]
 	for _, m := range providerCfg.Models {
 		if m.ID == modelID {
@@ -322,9 +322,9 @@ func (cfg *Config) ResolveModelTier(tier string) types.Model {
 	}
 }
 
-// resolveModelID returns the model ID string for the given tier,
+// ResolveModelID returns the model ID string for the given tier,
 // falling back to Model if the tier-specific value is not set.
-func (cfg *Config) resolveModelID(tier string) string {
+func (cfg *Config) ResolveModelID(tier string) string {
 	switch tier {
 	case ModelTierStrong:
 		if cfg.ModelStrong != "" {
