--- main.go
@@ -1,454 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log/slog"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"strings"
-	"syscall"
-	"time"
-
-	ucli "github.com/urfave/cli/v2"
-	"github.com/vaayne/anna/agent"
-	"github.com/vaayne/anna/agent/runner"
-	"github.com/vaayne/anna/agent/store"
-	"github.com/vaayne/anna/agent/tool"
-	"github.com/vaayne/anna/channel"
-	clicmd "github.com/vaayne/anna/channel/cli"
-	"github.com/vaayne/anna/channel/feishu"
-	"github.com/vaayne/anna/channel/qq"
-	"github.com/vaayne/anna/channel/telegram"
-	"github.com/vaayne/anna/cron"
-	"github.com/vaayne/anna/heartbeat"
-	"github.com/vaayne/anna/memory"
-	"github.com/vaayne/anna/skills"
-	"golang.org/x/sync/errgroup"
-)
-
-func main() {
-	slog.SetDefault(slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{
-		Level: slog.LevelDebug,
-	})))
-
-	app := newApp()
-
-	if err := app.Run(os.Args); err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-	}
-}
-
-func newApp() *ucli.App {
-	return &ucli.App{
-		Name:    "anna",
-		Usage:   "A local AI assistant",
-		Version: displayVersion(),
-		Commands: []*ucli.Command{
-			chatCommand(),
-			gatewayCommand(),
-			modelsCommand(),
-			skillsCommand(),
-			onboardCommand(),
-			versionCommand(),
-			upgradeCommand(),
-		},
-	}
-}
-
-func chatCommand() *ucli.Command {
-	return &ucli.Command{
-		Name:  "chat",
-		Usage: "Start interactive CLI chat",
-		Flags: []ucli.Flag{
-			&ucli.BoolFlag{
-				Name:  "stream",
-				Usage: "Read prompt from stdin and stream response to stdout",
-			},
-		},
-		Action: func(c *ucli.Context) error {
-			if !c.Bool("stream") {
-				if err := setupLogFile(); err != nil {
-					return fmt.Errorf("setup log file: %w", err)
-				}
-			}
-
-			s, err := setup(c.Context, false)
-			if err != nil {
-				return err
-			}
-			defer func() { _ = s.pool.Close() }()
-
-			if s.cronSvc != nil {
-				if err := s.cronSvc.Start(s.ctx); err != nil {
-					return fmt.Errorf("start cron: %w", err)
-				}
-				defer func() { _ = s.cronSvc.Stop() }()
-			}
-
-			if c.Bool("stream") {
-				return clicmd.RunStream(s.ctx, s.pool)
-			}
-			listFn := func() []channel.ModelOption { return collectModels(s.cfg) }
-			switchFn := modelSwitcher(s.cfg, s.pool, s.memStore, s.extraTools)
-			return clicmd.RunChat(s.ctx, s.pool, s.cfg.Provider, s.cfg.Model, listFn, switchFn)
-		},
-	}
-}
-
-func gatewayCommand() *ucli.Command {
-	return &ucli.Command{
-		Name:  "gateway",
-		Usage: "Start daemon services (Telegram, etc.) based on config",
-		Action: func(c *ucli.Context) error {
-			s, err := setup(c.Context, true)
-			if err != nil {
-				return err
-			}
-			defer func() { _ = s.pool.Close() }()
-
-			// Cron is started inside runGateway after notification wiring,
-			// so early-firing jobs already have the dispatcher callback.
-
-			listFn := func() []channel.ModelOption { return collectModels(s.cfg) }
-			switchFn := modelSwitcher(s.cfg, s.pool, s.memStore, s.extraTools)
-			return runGateway(s.ctx, s, listFn, switchFn)
-		},
-	}
-}
-
-type setupResult struct {
-	ctx        context.Context
-	cfg        *Config
-	pool       *agent.Pool
-	cronSvc    *cron.Service
-	heartbeat  *heartbeat.Service
-	memStore   *memory.Store
-	extraTools []tool.Tool
-	notifier   *channel.Dispatcher
-}
-
-func setup(parent context.Context, gateway bool) (*setupResult, error) {
-	cfg, err := LoadConfig()
-	if err != nil {
-		return nil, fmt.Errorf("loading config: %w", err)
-	}
-
-	ctx, cancel := signal.NotifyContext(parent, syscall.SIGINT, syscall.SIGTERM)
-	_ = cancel // cancel is deferred via the caller's lifecycle
-
-	// Create cron service and tool before the runner factory so the tool
-	// can be injected into the Go runner.
-	var cronSvc *cron.Service
-	var extraTools []tool.Tool
-	if cfg.Cron.CronEnabled() || (gateway && cfg.Heartbeat.IsEnabled()) {
-		cronSvc, err = cron.New(cfg.Cron.DataDir)
-		if err != nil {
-			return nil, fmt.Errorf("create cron service: %w", err)
-		}
-		if cfg.Cron.CronEnabled() {
-			extraTools = append(extraTools, cron.NewTool(cronSvc))
-		}
-	}
-
-	// Memory store + tool — always available.
-	memStore := memory.NewStore(cfg.MemoryPath())
-	extraTools = append(extraTools, memory.NewTool(memStore))
-
-	// Skills tool — always available.
-	cwd, _ := os.Getwd()
-	extraTools = append(extraTools, skills.NewTool(annaHome(), cfg.Workspace, cwd))
-
-	// Notification dispatcher + tool — backends are registered later in
-	// runGateway(). Only expose the tool in gateway mode where at least one
-	// enabled channel has enable_notify set to true.
-	dispatcher := channel.NewDispatcher()
-	if gateway && hasEnabledNotifyChannel(cfg) {
-		extraTools = append(extraTools, channel.NewNotifyTool(dispatcher))
-	}
-
-	idleTimeout := time.Duration(cfg.Runner.IdleTimeout) * time.Minute
-	factory, err := newRunnerFactory(cfg, memStore, extraTools)
-	if err != nil {
-		return nil, fmt.Errorf("create runner factory: %w", err)
-	}
-
-	opts := []agent.PoolOption{
-		agent.WithIdleTimeout(idleTimeout),
-		agent.WithCompaction(cfg.Runner.Compaction.WithDefaults()),
-		agent.WithDefaultModel(cfg.resolveModelID(ModelTierStrong)),
-		agent.WithFastModel(cfg.resolveModelID(ModelTierFast)),
-	}
-	sessionsPath := cfg.SessionsPath()
-	if sessionsPath != "" {
-		s, err := store.NewFileStore(sessionsPath, cwd)
-		if err != nil {
-			return nil, fmt.Errorf("create session store: %w", err)
-		}
-		opts = append(opts, agent.WithStore(s))
-		slog.Info("session persistence enabled", "dir", sessionsPath)
-	}
-
-	pool := agent.NewPool(factory, opts...)
-	go pool.StartReaper(ctx)
-
-	var heartbeatSvc *heartbeat.Service
-	if cfg.Heartbeat.IsEnabled() {
-		heartbeatSvc = heartbeat.New(heartbeat.Config{
-			File:      cfg.Heartbeat.FilePath(cfg.Workspace),
-			FastModel: cfg.resolveModelID(ModelTierFast),
-		}, func(ctx context.Context, sessionID, message, model string) <-chan runner.Event {
-			if model != "" {
-				return pool.Chat(ctx, sessionID, message, agent.WithModel(model))
-			}
-			return pool.Chat(ctx, sessionID, message)
-		}, dispatcher)
-	}
-
-	// Wire the cron callback now that pool exists.
-	if cronSvc != nil {
-		cronSvc.SetOnJob(func(ctx context.Context, job cron.Job) {
-			sessionID := job.SessionID()
-			msg := fmt.Sprintf("[Scheduled Task] %s\n\nInstruction: %s", job.Name, job.Message)
-			ch := pool.Chat(ctx, sessionID, msg)
-			for evt := range ch {
-				if evt.Err != nil {
-					slog.Error("cron job error", "job_id", job.ID, "error", evt.Err)
-				}
-			}
-		})
-	}
-
-	return &setupResult{
-		ctx:        ctx,
-		cfg:        cfg,
-		pool:       pool,
-		cronSvc:    cronSvc,
-		heartbeat:  heartbeatSvc,
-		memStore:   memStore,
-		extraTools: extraTools,
-		notifier:   dispatcher,
-	}, nil
-}
-
-func hasEnabledNotifyChannel(cfg *Config) bool {
-	return (cfg.Channels.Telegram.IsEnabled() && cfg.Channels.Telegram.IsNotifyEnabled()) ||
-		(cfg.Channels.QQ.IsEnabled() && cfg.Channels.QQ.IsNotifyEnabled()) ||
-		(cfg.Channels.Feishu.IsEnabled() && cfg.Channels.Feishu.IsNotifyEnabled())
-}
-
-func newRunnerFactory(cfg *Config, memStore *memory.Store, extraTools []tool.Tool) (runner.NewRunnerFunc, error) {
-	switch cfg.Runner.Type {
-	case "go":
-		providerCfg := cfg.Providers[cfg.Provider]
-		return func(ctx context.Context, model string) (runner.Runner, error) {
-			if model == "" {
-				model = cfg.Model
-			}
-			return runner.NewGoRunner(ctx, runner.GoRunnerConfig{
-				API:         cfg.Provider,
-				Model:       model,
-				APIKey:      providerCfg.APIKey,
-				Workspace:   cfg.Workspace,
-				AnnaHome:    annaHome(),
-				MemoryStore: memStore,
-				BaseURL:     providerCfg.BaseURL,
-				ExtraTools:  extraTools,
-			})
-		}, nil
-	default:
-		return nil, fmt.Errorf("unknown runner type: %q", cfg.Runner.Type)
-	}
-}
-
-// modelSwitcher returns a function that switches the pool's runner factory
-// to use a different provider/model combination.
-func modelSwitcher(cfg *Config, pool *agent.Pool, memStore *memory.Store, extraTools []tool.Tool) channel.ModelSwitchFunc {
-	return func(provider, model string) error {
-		cfg.Provider = provider
-		cfg.Model = model
-		factory, err := newRunnerFactory(cfg, memStore, extraTools)
-		if err != nil {
-			return err
-		}
-		pool.SetFactory(factory)
-		pool.SetDefaultModel(model)
-		if err := SaveModelSelection(cfg.Workspace, provider, model); err != nil {
-			slog.Warn("failed to persist model selection", "error", err)
-		}
-		return nil
-	}
-}
-
-func setupLogFile() error {
-	logPath := filepath.Join(annaHome(), "anna.log")
-	if err := os.MkdirAll(filepath.Dir(logPath), 0o755); err != nil {
-		return err
-	}
-	f, err := os.OpenFile(logPath, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0o644)
-	if err != nil {
-		return err
-	}
-	slog.SetDefault(slog.New(slog.NewTextHandler(f, &slog.HandlerOptions{
-		Level: slog.LevelDebug,
-	})))
-	return nil
-}
-
-func runGateway(ctx context.Context, s *setupResult, listFn channel.ModelListFunc, switchFn channel.ModelSwitchFunc) error {
-	g, gctx := errgroup.WithContext(ctx)
-	var channels []channel.Channel
-
-	// --- Telegram ---
-	tg := s.cfg.Channels.Telegram
-	if tg.IsEnabled() && tg.Token != "" {
-		slog.Info("starting telegram bot")
-
-		tgBot, err := telegram.New(telegram.Config{
-			Token:      tg.Token,
-			NotifyChat: tg.NotifyChat,
-			ChannelID:  tg.ChannelID,
-			GroupMode:  tg.GroupMode,
-			AllowedIDs: tg.AllowedIDs,
-		}, s.pool, listFn, switchFn)
-		if err != nil {
-			return fmt.Errorf("create telegram bot: %w", err)
-		}
-
-		defaultChat := tg.NotifyChat
-		if defaultChat == "" {
-			defaultChat = tg.ChannelID
-		}
-		channels = append(channels, tgBot)
-		if tg.IsEnabled() && tg.IsNotifyEnabled() {
-			s.notifier.Register(tgBot, defaultChat)
-		}
-	}
-
-	// --- QQ ---
-	qqCfg := s.cfg.Channels.QQ
-	if qqCfg.IsEnabled() && qqCfg.AppID != "" && qqCfg.AppSecret != "" {
-		slog.Info("starting qq bot")
-
-		qqBot, err := qq.New(qq.Config{
-			AppID:      qqCfg.AppID,
-			AppSecret:  qqCfg.AppSecret,
-			GroupMode:  qqCfg.GroupMode,
-			AllowedIDs: qqCfg.AllowedIDs,
-		}, s.pool, listFn, switchFn)
-		if err != nil {
-			return fmt.Errorf("create qq bot: %w", err)
-		}
-
-		channels = append(channels, qqBot)
-		if qqCfg.IsEnabled() && qqCfg.IsNotifyEnabled() {
-			s.notifier.Register(qqBot, "")
-		}
-	}
-
-	// --- Feishu ---
-	fsCfg := s.cfg.Channels.Feishu
-	if fsCfg.IsEnabled() && fsCfg.AppID != "" && fsCfg.AppSecret != "" {
-		slog.Info("starting feishu bot")
-
-		fsBot, err := feishu.New(feishu.Config{
-			AppID:             fsCfg.AppID,
-			AppSecret:         fsCfg.AppSecret,
-			EncryptKey:        fsCfg.EncryptKey,
-			VerificationToken: fsCfg.VerificationToken,
-			NotifyChat:        fsCfg.NotifyChat,
-			GroupMode:         fsCfg.GroupMode,
-			AllowedIDs:        fsCfg.AllowedIDs,
-		}, s.pool, listFn, switchFn)
-		if err != nil {
-			return fmt.Errorf("create feishu bot: %w", err)
-		}
-
-		channels = append(channels, fsBot)
-		if fsCfg.IsEnabled() && fsCfg.IsNotifyEnabled() {
-			s.notifier.Register(fsBot, fsCfg.NotifyChat)
-		}
-	}
-
-	if len(channels) == 0 {
-		return fmt.Errorf("no gateway services configured. Check %s", configPath())
-	}
-
-	if len(s.notifier.Channels()) == 0 {
-		slog.Warn("no enabled channels have enable_notify set to true; cron results and heartbeat notifications will not be delivered")
-	}
-
-	// Start all channels.
-	for _, ch := range channels {
-		g.Go(func() error {
-			if err := ch.Start(gctx); err != nil && gctx.Err() == nil {
-				return fmt.Errorf("%s: %w", ch.Name(), err)
-			}
-			return nil
-		})
-	}
-
-	// Wire cron notifications and start the scheduler AFTER channels
-	// are registered, so early-firing jobs already use the dispatcher.
-	if s.cronSvc != nil {
-		if s.cfg.Cron.CronEnabled() {
-			wireCronNotifier(s.cronSvc, s.pool, s.notifier)
-			if err := s.cronSvc.Start(ctx); err != nil {
-				return fmt.Errorf("start cron: %w", err)
-			}
-		} else {
-			if err := s.cronSvc.StartEphemeral(ctx); err != nil {
-				return fmt.Errorf("start shared scheduler: %w", err)
-			}
-		}
-		defer func() { _ = s.cronSvc.Stop() }()
-	}
-
-	if s.heartbeat != nil {
-		if s.cronSvc == nil {
-			return fmt.Errorf("heartbeat requires the shared scheduler")
-		}
-		if err := scheduleHeartbeat(ctx, s.cronSvc, s.heartbeat, s.cfg.Heartbeat.Interval()); err != nil {
-			return fmt.Errorf("schedule heartbeat: %w", err)
-		}
-	}
-
-	err := g.Wait()
-	slog.Info("gateway stopped")
-	return err
-}
-
-// wireCronNotifier overrides the cron callback to collect the agent response
-// and broadcast it via the notification dispatcher.
-func wireCronNotifier(cronSvc *cron.Service, pool *agent.Pool, dispatcher *channel.Dispatcher) {
-	cronSvc.SetOnJob(func(ctx context.Context, job cron.Job) {
-		sessionID := job.SessionID()
-		msg := fmt.Sprintf("[Scheduled Task] %s\n\nInstruction: %s", job.Name, job.Message)
-		var result strings.Builder
-		for evt := range pool.Chat(ctx, sessionID, msg) {
-			if evt.Err != nil {
-				slog.Error("cron job error", "job_id", job.ID, "error", evt.Err)
-			}
-			if evt.Text != "" {
-				result.WriteString(evt.Text)
-			}
-		}
-		if result.Len() > 0 {
-			text := fmt.Sprintf("*%s*\n\n%s", job.Name, result.String())
-			if err := dispatcher.Notify(ctx, channel.Notification{Text: text}); err != nil {
-				slog.Error("cron notification failed", "job_id", job.ID, "error", err)
-			}
-		}
-	})
-}
-
-func scheduleHeartbeat(ctx context.Context, cronSvc *cron.Service, heartbeatSvc *heartbeat.Service, every string) error {
-	slog.Info("starting heartbeat", "every", every)
-	return cronSvc.ScheduleEvery(ctx, every, func(ctx context.Context) {
-		if err := heartbeatSvc.Poll(ctx); err != nil {
-			slog.Error("heartbeat poll failed", "error", err)
-		}
-	})
-}
