+++ mise.toml
@@ -12,7 +12,7 @@ BINARY_NAME = "anna"
 
 [tasks.build]
 description = "Build anna binary"
-run = "VERSION=${VERSION:-$(git describe --tags --always --dirty 2>/dev/null || echo dev)} && go build -ldflags=\"-X main.version=$VERSION\" -o bin/$BINARY_NAME ."
+run = "VERSION=${VERSION:-$(git describe --tags --always --dirty 2>/dev/null || echo dev)} && go build -ldflags=\"-X main.version=$VERSION\" -o bin/$BINARY_NAME ./cmd/anna"
 
 [tasks.clean]
 description = "Remove build artifacts"
